+++ runners/flowise/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @proompter/runner-flowise
 
+## 0.0.8
+
+### Patch Changes
+
+- remove mkdir calls
+
 ## 0.0.7
 
 ### Patch Changes