+++ server/next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @proompter/server-next
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @proompter/runner-flowise@0.0.8
+
 ## 0.0.10
 
 ### Patch Changes